pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vault-mcp"
version = "0.2.5"
description = "Personal knowledge vault MCP server for Claude Desktop"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "python-frontmatter",
    "pydantic>=2.0",
    "python-dotenv",
    "pyyaml",
    "networkx>=3.0",
    "python-louvain>=0.16",
]

[project.scripts]
vault-mcp = "vault_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vault_mcp"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
